# openapi.yaml
openapi: 3.0.3
info:
  title: PennyWise Finance Tracker API
  description: |
    Personal finance tracker with income/expense tracking, budgets, analytics,
    receipt OCR, recurring transactions, assets/net worth, currency conversion,
    CSV import/export and offline PWA support.
  version: 1.0.0
  contact:
    name: PennyWise Team
servers:
  - url: /api
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /income:
    get:
      summary: List income records
      parameters:
        - { name: from, in: query, schema: { type: string } }
        - { name: to, in: query, schema: { type: string } }
        - { name: category, in: query, schema: { type: string } }
        - { name: month, in: query, schema: { type: string } }
      responses:
        '200':
          description: Income list
    post:
      summary: Create income
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount, source, category, date]
              properties:
                amount: { type: number, exclusiveMinimum: 0 }
                source: { type: string }
                category: { type: string }
                date: { type: string, format: date }
                is_recurring: { type: boolean }
                currency: { type: string, pattern: '^[A-Z]{3}$', default: USD }
                notes: { type: string }
      responses:
        '201':
          description: Created income record
  /expenses:
    get:
      summary: List expenses
      parameters:
        - { name: from, in: query, schema: { type: string } }
        - { name: to, in: query, schema: { type: string } }
        - { name: category, in: query, schema: { type: string } }
        - { name: merchant, in: query, schema: { type: string } }
      responses:
        '200':
          description: Expense list
    post:
      summary: Create expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount, merchant, category, date]
              properties:
                amount: { type: number, exclusiveMinimum: 0 }
                merchant: { type: string }
                category: { type: string }
                date: { type: string, format: date }
                currency: { type: string, pattern: '^[A-Z]{3}$', default: USD }
                receipt_path: { type: string }
                project_id: { type: integer }
                notes: { type: string }
      responses:
        '201':
          description: Created expense
...
  /projects:
    get:
      summary: List project budgets
    post:
      summary: Create a project budget
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name, total_budget, start_date]
              properties:
                name: { type: string }
                total_budget: { type: number, exclusiveMinimum: 0 }
                start_date: { type: string, format: date }
                end_date: { type: string, format: date }
      responses:
        '201':
          description: Created project
  /projects/{id}/expenses:
    get:
      summary: List expenses linked to a project
    post:
      summary: Link an expense to a project (increments spent)
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [expense_id]
              properties:
                expense_id: { type: integer }
      responses:
        '201':
          description: Expense linked
        '409':
          description: Already linked
...
  /recurring:
    get:
      summary: List recurring rules (also processes due rules)
    post:
      summary: Create a recurring transaction
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [type, amount, description, frequency, next_date]
              properties:
                type: { type: string, enum: [income, expense] }
                amount: { type: number }
                description: { type: string }
                frequency: { type: string, enum: [daily, weekly, monthly, yearly] }
                next_date: { type: string, format: date }
  /recurring/upcoming:
    get:
      summary: List upcoming recurring payments within N days
      parameters:
        - { name: days, in: query, schema: { type: integer, default: 14 } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: pennywise_token